refactor(bills): move bill decorators + modals into :shared:bills
Phase A of hosting bills at the app root: relocate the below-bill decorators
(ScannableDecorator/PayableDecorator/TipCardDecorator) and their modals
(BillManagementOptions, ReceivedFundsConfirmation, TipUserModal) out of
:features:scanner into :shared:bills, so the bill overlay can later be hosted
outside the scanner. Pure move — no behavior change.

- ScannableDecorator + ScannableDecoratorContext become public (the scanner
  now consumes them cross-module); the concrete decorators/modals stay
  internal to :shared:bills.
- :shared:bills gains a :shared:session dependency (for LocalSessionController).
  The 7 string resources resolve transitively (:core + :ui:resources), so none
  were moved.
- ScannableContainer repointed to the new com.flipcash.app.bills.decor package.
