+++ apps/flipcash/shared/bills/build.gradle.kts
@@ -12,6 +12,7 @@ dependencies {
 
     implementation(project(":libs:messaging"))
     implementation(project(":apps:flipcash:shared:common-ui"))
+    implementation(project(":apps:flipcash:shared:session"))
 
     implementation(libs.androidx.datastore)
     implementation(libs.bundles.haze)
